Extract common Service functionality for HTTP based stream sinks.
