+++ src/sinks/http/request_builder.rs
@@ -4,9 +4,9 @@ use std::io;
 
 use bytes::Bytes;
 
-use crate::sinks::prelude::*;
+use crate::sinks::{prelude::*, util::http_service::HttpRequest};
 
-use super::{encoder::HttpEncoder, service::HttpRequest};
+use super::encoder::HttpEncoder;
 
 pub(super) struct HttpRequestBuilder {
     pub(super) encoder: HttpEncoder,
