+++ src/sinks/util/http_service/retry.rs
@@ -1,0 +1,38 @@
+use http::StatusCode;
+
+use crate::{
+    http::HttpError,
+    sinks::util::retries::{RetryAction, RetryLogic},
+};
+
+use super::service::HttpResponse;
+
+/// `RetryLogic` implementation for use in HTTP based stream sinks.
+#[derive(Debug, Default, Clone)]
+pub struct HttpRetryLogic;
+
+impl RetryLogic for HttpRetryLogic {
+    type Error = HttpError;
+    type Response = HttpResponse;
+
+    fn is_retriable_error(&self, _error: &Self::Error) -> bool {
+        true
+    }
+
+    fn should_retry_response(&self, response: &Self::Response) -> RetryAction {
+        let response = &response.http_response;
+        let status = response.status();
+
+        match status {
+            StatusCode::TOO_MANY_REQUESTS => RetryAction::Retry("too many requests".into()),
+            StatusCode::NOT_IMPLEMENTED => {
+                RetryAction::DontRetry("endpoint not implemented".into())
+            }
+            _ if status.is_server_error() => RetryAction::Retry(
+                format!("{}: {}", status, String::from_utf8_lossy(response.body())).into(),
+            ),
+            _ if status.is_success() => RetryAction::Successful,
+            _ => RetryAction::DontRetry(format!("response status: {}", status).into()),
+        }
+    }
+}
